requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.37.0
pydantic==2.10.3
requests==2.32.3
python-dotenv==1.0.0
httpx==0.27.2
